feat: add solve to modules task
